Created a script that notifies the area admin a few days in advance about the jobs from the area. The script sends the contact details, emails, and previous experience of the job participants.
It should be added to a cronjob for the given days in advance. For example if the jobs are done Tuesday and Thursday
and the notification should be sent 2 days in advance, one must add a cronjob:
Sundays and Tuesdays at 8:00 am with the command

python manage.py upcoming_jobs --area Verteilen --days 2

This will send an email to the area admin with a summary of the jobs.